add a new e2e test on fscore 0.5

+++ tests/end2end-tests/graphstorm-nc/mgpu_test.sh
@@ -883,4 +883,20 @@ fi
 
 rm /tmp/train_log.txt
 
+echo "**************dataset: MovieLens: NC, RGCN layer: 1, node feat: fixed, eval metrics: fscore_at_0.5"
+python3 -m graphstorm.run.gs_node_classification --workspace $GS_HOME/training_scripts/gsgnn_np/ --num-trainers $NUM_TRAINERS --num-servers 1 --num-samplers 0 --part-config /data/movielen_100k_ef_nc_ec_train_val_1p_4t/movie-lens-100k.json --ip-config ip_list.txt --ssh-port 2222 --cf ml_nc.yaml --node-feat-name user:feat --batch-size 64 --save-model-path /data/gsgnn_nc_ml_ef/model/ --save-model-frequency 5 --eval-frequency 1  --num-epochs 2 --logging-file /tmp/train_log.txt --backend nccl --model-encoder-type rgcn --eval-metric fscore_at_0.5
+
+error_and_exit $?
+
+fscore_eval_cnts=$(grep "fscore_at_0.5:" /tmp/train_log.txt | wc -l)
+
+if [ $fscore_eval_cnts < 1 ]
+then
+    echo "The number of fscore_at_0.5 should be greater than 0, but got $fscore_eval_cnts."
+    exit -1
+fi
+
+rm /tmp/train_log.txt
+
+
 rm -fr /tmp/*
